etl/requirements.txt
# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-celery==3.5.0
apache-airflow-providers-redis==3.5.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-amazon==8.15.0
apache-airflow-providers-elasticsearch==5.3.0

# Celery
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Celery monitoring

# Database
psycopg2-binary==2.9.9
SQLAlchemy==1.4.51

# Elasticsearch
elasticsearch==8.11.1

# AWS S3
boto3==1.34.10
botocore==1.34.10

# Video processing
Pillow==10.2.0  # Image processing

# HTTP
requests==2.31.0
httpx==0.26.0

# Data processing
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Note: FFmpeg must be installed separately in the Docker image
# apt-get install -y ffmpeg
